requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-community
langchain-openai
sqlite3
json
hashlib
datetime